requirements.txt
click==8.1.3
Flask==2.2.3
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
python-dotenv==0.17.1
SQLAlchemy==2.0.4
typing_extensions==4.4.0
Werkzeug==2.2.3
zipp==3.15.0
